+++ VIPRA/Extendable/pedestrianset/pedestrian_set.hpp
@@ -23,7 +23,7 @@ class PedestrianSet {
 
   virtual void updateState(std::shared_ptr<VIPRA::State>) = 0;
 
-  virtual int                       getNumPedestrians() const noexcept = 0;
+  virtual size_t                    getNumPedestrians() const noexcept = 0;
   virtual const VIPRA::f3dVec&      getPedestrianCoordinates() const noexcept = 0;
   virtual const VIPRA::f3dVec&      getVelocities() const noexcept = 0;
   virtual const std::vector<float>& getSpeeds() const noexcept = 0;
